Move registration provider impls
