+++ Forge/src/main/java/potionstudios/byg/registration/ForgeBlockFactory.java
@@ -1,4 +1,4 @@
-package potionstudios.byg;
+package potionstudios.byg.registration;
 
 import com.google.auto.service.AutoService;
 import net.minecraft.world.level.block.Block;
@@ -9,9 +9,9 @@ import potionstudios.byg.common.block.BYGBlockFactory;
 import java.util.function.Supplier;
 
 @AutoService(BYGBlockFactory.class)
-public class FabricBlockFactory implements BYGBlockFactory {
+public class ForgeBlockFactory implements BYGBlockFactory {
     @Override
     public FlowerPotBlock createPottedBlock(Supplier<? extends Block> block, BlockBehaviour.Properties behaviour) {
-        return new FlowerPotBlock(block.get(), behaviour);
+        return new FlowerPotBlock(null, block, behaviour);
     }
 }
